refactor: remove unused css class
