+++ src/components/TheSettingsMenu.vue
@@ -80,6 +80,7 @@ import SettingsDashboardTab from '@/components/settings/SettingsDashboardTab.vue
 import SettingsGCodeViewerTab from '@/components/settings/SettingsGCodeViewerTab.vue'
 import SettingsEditorTab from '@/components/settings/SettingsEditorTab.vue'
 import SettingsTimelapseTab from '@/components/settings/SettingsTimelapseTab.vue'
+import SettingsNavigationTab from '@/components/settings/SettingsNavigationTab.vue'
 
 import Panel from '@/components/ui/Panel.vue'
 import {
@@ -98,6 +99,7 @@ import {
     mdiVideo3d,
     mdiWebcam,
     mdiDipSwitch,
+    mdiMenu,
 } from '@mdi/js'
 import SettingsMiscellaneousTab from '@/components/settings/SettingsMiscellaneousTab.vue'
 @Component({
@@ -116,6 +118,7 @@ import SettingsMiscellaneousTab from '@/components/settings/SettingsMiscellaneou
         SettingsEditorTab,
         SettingsTimelapseTab,
         SettingsMiscellaneousTab,
+        SettingsNavigationTab,
     },
 })
 export default class TheSettingsMenu extends Mixins(BaseMixin) {
@@ -193,6 +196,11 @@ export default class TheSettingsMenu extends Mixins(BaseMixin) {
                 icon: mdiDipSwitch,
                 name: 'miscellaneous',
                 title: this.$t('Settings.MiscellaneousTab.Miscellaneous'),
+            },
+            {
+                icon: mdiMenu,
+                name: 'navigation',
+                title: this.$t('Settings.NavigationTab.Navigation'),
             },
         ]
