+++ src/components/TheSidebar.vue
@@ -45,16 +45,14 @@
 
 <script lang="ts">
 import Component from 'vue-class-component'
-import routes, { AppRoute } from '@/routes'
-import { Mixins, Watch } from 'vue-property-decorator'
+import { Mixins } from 'vue-property-decorator'
 import BaseMixin from '@/components/mixins/base'
-import { PrinterStateKlipperConfig } from '@/store/printer/types'
 import TheSelectPrinterDialog from '@/components/TheSelectPrinterDialog.vue'
 import AboutDialog from '@/components/dialogs/AboutDialog.vue'
 import { navigationWidth, topbarHeight } from '@/store/variables'
 import MainsailLogo from '@/components/ui/MainsailLogo.vue'
-import { mdiViewDashboardOutline, mdiLinkVariant } from '@mdi/js'
-import SidebarItem, { NaviPoint } from '@/components/ui/SidebarItem.vue'
+import SidebarItem from '@/components/ui/SidebarItem.vue'
+import NavigationMixin from '@/components/mixins/navigation'
 
 @Component({
     components: {
@@ -64,11 +62,9 @@ import SidebarItem, { NaviPoint } from '@/components/ui/SidebarItem.vue'
         MainsailLogo,
     },
 })
-export default class TheSidebar extends Mixins(BaseMixin) {
+export default class TheSidebar extends Mixins(NavigationMixin, BaseMixin) {
     navigationWidth = navigationWidth
     topbarHeight = topbarHeight
-
-    private customNaviLinks: NaviPoint[] = []
 
     get naviDrawer(): boolean {
         return this.$store.state.naviDrawer
@@ -86,68 +82,8 @@ export default class TheSidebar extends Mixins(BaseMixin) {
         return this.$store.getters['files/getSidebarBackground']
     }
 
-    get routesNaviPoints(): NaviPoint[] {
-        const points: NaviPoint[] = []
-
-        if (this.countPrinters) {
-            points.push({
-                title: this.$t('App.Printers'),
-                icon: mdiViewDashboardOutline,
-                to: '/allPrinters',
-                position: 0,
-            } as NaviPoint)
-        }
-
-        routes
-            .filter((element) => {
-                return element.showInNavi && this.showInNavi(element)
-            })
-            .forEach((element) => {
-                points.push({
-                    title: this.$t(`Router.${element.title}`),
-                    icon: element.icon,
-                    to: element.path,
-                    position: element.position ?? 999,
-                } as NaviPoint)
-            })
-
-        if (this.customNaviLinks.length) {
-            points.push(...this.customNaviLinks)
-        }
-
-        return points
-    }
-
-    get naviPoints(): NaviPoint[] {
-        return this.routesNaviPoints.sort((a, b) => a.position - b.position)
-    }
-
-    get klippy_state(): string {
-        return this.$store.state.server.klippy_state
-    }
-
-    get boolNaviWebcam(): boolean {
-        return this.$store.state.gui.uiSettings.boolWebcamNavi
-    }
-
-    get moonrakerComponents(): string[] {
-        return this.$store.state.server.components
-    }
-
-    get registeredDirectories(): string[] {
-        return this.$store.state.server.registered_directories
-    }
-
-    get klipperConfigfileSettings(): PrinterStateKlipperConfig[] {
-        return this.$store.state.printer.configfile?.settings ?? {}
-    }
-
     get currentPage(): string {
         return this.$route.fullPath
-    }
-
-    get countPrinters() {
-        return this.$store.getters['farm/countPrinters']
     }
 
     get boolNaviTemp(): boolean {
@@ -166,14 +102,6 @@ export default class TheSidebar extends Mixins(BaseMixin) {
 
     get sidebarLogo(): string {
         return this.$store.getters['files/getSidebarLogo']
-    }
-
-    get sidebarNaviFile(): string {
-        const file = this.$store.getters['files/getCustomNaviPoints']
-
-        window.console.log(file)
-
-        return file
     }
 
     get logoColor(): string {
@@ -204,40 +132,6 @@ export default class TheSidebar extends Mixins(BaseMixin) {
         }
 
         return output
-    }
-
-    @Watch('sidebarNaviFile', { immediate: true })
-    async sidebarNaviFileChanged(newVal: string) {
-        this.customNaviLinks = []
-
-        const content = await fetch(newVal)
-            .then((res) => res.json())
-            .catch((err) => {
-                window.console.error('Unable to parse .theme/navi.json.')
-                throw err
-            })
-
-        content.forEach((item: NaviPoint) => {
-            this.customNaviLinks.push({
-                title: item.title ?? 'Unknown',
-                icon: item.icon ?? mdiLinkVariant,
-                href: item.href ?? '#',
-                target: item.target ?? undefined,
-                position: item.position ?? 999,
-            } as NaviPoint)
-        })
-    }
-
-    showInNavi(route: AppRoute): boolean {
-        if (['shutdown', 'error', 'disconnected'].includes(this.klippy_state) && !route.alwaysShow) return false
-        else if (route.title === 'Webcam' && !this.boolNaviWebcam) return false
-        else if (route.moonrakerComponent && !this.moonrakerComponents.includes(route.moonrakerComponent)) return false
-        else if (route.registeredDirectory && !this.registeredDirectories.includes(route.registeredDirectory))
-            return false
-        else if (route.klipperComponent && !(route.klipperComponent in this.klipperConfigfileSettings)) return false
-        else if (route.klipperIsConnected && !this.klippyIsConnected) return false
-
-        return true
     }
 
     mounted() {
